// src/config/schema.ts
export interface ScopeConfig {
  writable?: string[];
  readable?: string[];
}